backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
pydantic==2.6.4
pydantic-settings==2.2.1
httpx==0.26.0 